requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
petl
requests
python-dateutil
django-bootstrap4

# Dev-dependencies
django-extensions
ipdb
pyreadline
ipython
